Added first draft of scripts to start locators, leads and servers on multiple machines. These are based on spark scripts 1. start-snappy-all starts locators leads and servers 2. locators.sh start/stops/queries status of locator on all the nodes that are specified in conf/locators file 3. servers.sh start/stops/queries status of servers on all the nodes that are specified in conf/servers file 4. leads.sh start/stops/queries status of leads on all the nodes that are specified in conf/leads file
Both snappy-env.sh and spark-env.sh will be read for environment settings.
Both snappy-config.sh abd spark-config.sh will be read for config settings.

There is an IllegalAccessError thrown while calling functions in parent class which is in java from scala. Added a overriden function in GfxdLauncherOverrides.scala to overcome this error.

Updated the build-script to copy the script files.