My analysis of the issue: AkkaRpcEnv.setupEndpoint creates a AkkaRpcEndpointRef. This AkkaRpcEndpointRef has a transient val _actorRef which is assigned to a lazy val actorRef. Now the actorRef gets serialized while _actorRef does not. There is a possibility that AkkaRpcEndpointRef is serialized before lazy val actorRef was materialized. And hence this may result in the exception that we are seeing in the driver side.
However fixing this would need more  understanding of the code. Hence I am fixing it in our code rather than in Spark by retrying. After an Executor fails to start, retry starting the executor after a wait for 3 seconds. I don't limit the number of retries because this issue is expected to go away after retries.