changing char datatype to varchar
Changing the data type from char to varchar as spark RDD doesnot differentiate b/w char and
varchar. Above all in the in test for spark all the data is cached in memmory so no db interface comes to the picture
whereas gemxd treats char(fixed size and appends spaces at end to fill the length) and varchar differently and that is causing the
row count mismatch between spark and gemxd.