We had made a decision to ship aqp classes in assembly jar but not to ship aqp code.
So, I have added code in the build.gradle that checks if the aqp project exists, if yes, it adds the aqp classes to the assembly jar. If no, nothing is done for now. Later we can upload the aqp jar somewhere and this code will fetch that jar and make it part of assembly jar.