added the docs of generating all java/scala docs for all the submodules.
what is remaining is to add dependency on the target 'product' so that snappy-store docs can also get built
and also copy the snappy-docs to the main product docs dir. An index.html has been added which has a link to
all the submodules on it. Need to add a page title also to the top level index.html

Use this to generate the docs: ./gradlew :snappy_tools_2.10:docs
and open the main index.html file in the root projects build dir named docs.

file:///home/kneeraj/snappy/snappy-commons/build-artifacts/scala-2.10/docs/index.html