Issue : While saving a dataframe to row table if a varchar field contains null it gives an exception saying wrong access to CLOb type
Analysis :
While saving a dataframe in row store we use code generation to set PrepearedStatement values.
This code ignores the fact that row table might contain an actual VARCHAR and not a CLOB types as is done in column tables.

Fix : Passed field.metadata while getting JDBC types. It ensures if a VARCHAR is there in underlying storage it will return a VARCHAR.