snappy cleanup handled during SparkContext stop currently SnappyContext.stop takes care of all the snappy clean up activities  and it also stops the sparkContext. Now all the clean up  happens during the sparkContext.stop
All the tests are also changed to use the sc.stop instead
of SparkContext.stop

default time to run the MemoryAnalytics service is changed
to 1 second for the tests

also adding the missing UI files to display the Snappy tab.

All the methods from GlobalSnappyInit are moved to SnappyContext
as it does not make sense to keep two singleton for same purpose.
just keeping the AQP related stuff in the GlobalSnappyInit
which can be removed later.