Removed some of the thread unsafe code from SnappyStoreHiveCatalog. (#259)
* Removed some of the thread unsafe code from SnappyStoreHiveCatalog.
a) tempTable structures made concurrent hash map.
b) SessionState in hive client is a thread local variable and need to be set for every thread. HiveContext does it by HiveQL. The same need to be done for SnappyContext. As SnappyContext also has helper APIs, I have taken care of setting a SessionState inside the catalog.
c) If we do this probably we dont need to remove SnappyContext.getOrCreate();

This code might need some rework once we merge Spark2.0.

* Added a test to check SnappyContext concurrency.

* As long as threads using same IsolatedClientLoader, they should
use the cached Hive client and not create a new one. we will be setting the cached client as
soon as the ClientWrapper is initialized, so that any thread accessing Hive client after
that will get the cached client. This change will be redundant after Spark 2.0 merge. But
for the time being it will avoid ThreadLocal access to set SessionState.

* Corrected the spelling mistake in test file.