SnappyUnifiedMemoryManager (#260)
Added a SnappyUnifiedMemoryManager. The idea is that when there is request for execution or storage memory, critical up and eviction up events are checked. If they are set, we try to free the memory cached by Spark rdds by calling memoryStore.evictBlocksToFreeSpace. If if evictBlocksToFreeSpace has been invoked from acquireStorageMemory, return false if  evictBlocksToFreeSpace fails and evictionUp is still true. If evictBlocksToFreeSpace has been invoked from acquireExecutionMemory then only fail if criticalUp is still true after evictBlocksToFreeSpace. 

With 1.6, Spark is accounting memory for execution and storage pools. When one pool grows, it can borrow from another pool but keeps track of it. This accounting is lost 
with Snappy and we need to fix it with a new resource manager.  

Tested it using the existing SnappyResourceEventsDUnitTest.