First changes for the SQL parser
DDL Parser
Things to make it work right now:
Parser has changed with 2.0. We will copy the old Spark DDL parser from 1.6 into Snappy and make it work.

Things to do before GA:
We will move to the new Spark parser and remove the 1.6 parser.

SQL Parser
Things to make it work right now:
We are anyway using a different parser. Nothing to do here.

Things to do before GA:
Move to antlr based new Spark parser if the issues because of which we introduced our parser are fixed. Else, we need to define how we would like to add new extensions that are already part of new Spark parser.