Ensuring that Spark Uncaught exceptions are handled in the Snappy side and do not cause a system.exit (#342)
SnappyUncaughtExceptionHandler is now set as the uncaught exception handler in the executor. SnappyUncaughtExceptionHandler exits the executor component unlike the SparkUncaughtExceptionHandler which does a System.exit.

While exiting the executor, check if snappy store is shutting down. If yes, there is no need to restart the executor.

Patch testing

Added a test that tests restarting of executors because of an uncaught thread exception.
Added a test that tests restarting of executors because of an executor going down because of a fatal error.