Snap 1087 (#404)
This pull request maintains stats (which include lower bound, upper bound, null count etc) for every column. And then uses the upper bound and lower bound values of columns to filter out the cached batches. This will be a perf enhancement for the queries which filters extensively.

To maintain the stats, enabled the stat generation code that was already there and maintained stats as part of the cached batch.

To filter out the stats, following things were done:

    Take the filter predicates and the stats schema from the relation to the scan level so that they can be utilized. For this a class StatsPredicate was created.
    Compile the filter predicates to form a SpecificPredicate class on every executor. The filter predicate creation and compilation was picked up from InMemoryTableScanExec.
    During iteration make an UnsafeRow out of the stats bytes and then apply the predicate on this UnsafeRow.
    Added stats for cached batches seen and skipped

Currently this optimization does not apply for split mode, for sample tables and for off heap iterators. These will be fixed in next few PRs.

Precheckin was run. Added a test case that reads the stats and verifies the number of cached batches scanned.