Adding cache for optimized local join map
The cache is created when the first partition asks for it and maintained
only till the last partition references it. This means that the map
can potentially get re-created multiple times if all partitions did
not get scheduled to an executor in one shot.

This is acceptable given that this avoids the complications of
invalidating the cache.