Executorcatalog (#500)
Changes proposed in this pull request

Before this checkin, the table meta data information on executors was supplied by the StoreInitRDD. This is not correct because if a table is created from SparkApp and then data is inserted into it, it fails to age the data to column buffer as the table's metadata is not available. Also, in HA conditions, this could create problems. The table metadata is anyway available on the executors inside hive catalog. Hence added code to read the hive meta data from the catalog on the executor itself.

Secondly, cached batch size and usecompression were read from sqlconf. This is incorrect because the behaviour is confusing when the data is being inserted irom a spark app. Hence made these two parameters part of table metadata. They can be specified as part of the table properties. cached batch size if not specified as table properties, we try to read it from spark conf. This was done to make the exisiting test cases work. If it is not part of spark conf then default of 10000 is being used.

After these two changes, store init rdd was not needed for snappydata data core but it was needed for aqp. So moved it to aqp.

When a table is created using spark way of saveAsTable instead of createTable, the dd entry was created, data was inserted and then hive catalog entry is created. So, during data insertion the hive meta data was not found on the executors. To fix this , moved the hive catalog registration after the relation is created inside the createRelation method. Also since, createRelation can be called multiple times for a single table, avoided registration in hive catalog multiple times.

Fixed the tests to set the cached batch sizes in the sparkconf instead of sqlconf.
Patch testing

Clean precheckin. Will be adding a test to verify the hive meta data on executors in embedded and connector modes.
Other PRs


SnappyDataInc/snappy-store#146