Code generated plan for row table to column table insert, row table insert
Refactored ColumnInsertExec into separate ColumnWriter singleton to enable using from
StoreCallbacksImpl when moving from row buffer to column table.

Also added RowInsertExec for efficient row table inserts. Now also used for the case
when the last batch in column table is very small, then it is decompressed and inserted into row buffer.

Use above plans uniformly for base JDBCMutableRelation and JDBCAppendableRelation too.

Added full support for complex types in inserts with updated ComplexTypeSerializerImpl to use the same.
New SerializedRow, SerializedArray and SerializedMap classes for complex types that respect
the platform endianness (use little-endian on all platforms since that is the more common platform now)

Added a "skipBytes" mechanism in Serialized* classes above to enable 8-byte word aligned reads.
For example when writing 4-byte size at the start of a SerializedRow, the skipBytes will be
set to 4 and the baseOffset will point to before the size and those 4 bytes will be subsumed
in the nullMask at the start (with the skipBytes taking care of bytes to not consider as nullMask)

Set thrift to true by default when booting from SnappyData

Fixing number of test failures many of which are also present in master: ExecutionEngineArbiterDUnitTest,
       ColumnTableDUnitTest, QueryRoutingDUnitTest.testGemXDURL, Split*ClusterDUnitTest,
       ColumnTableInternalValidationTest, ColumnTableTest etc