[SNAP-1398] Update janino version to latest 3.0.x
This works around some of the limitations of older janino versions. But the fundamental
JVM limitation of 64K method bytecode still remains and hits for larger number of columns
which needs more work to fix to split generated code of scans and inserts.

Also updated snappy-spark link for the updated janino version.