Use local variables for simple aggregates with column scans
Using local variables (and updating member variables at the end before return)
squeezes a further 10% improvement in the single threaded ColumnCacheBenchmark
test making it nearly on-par with custom hand-code array reads+average.

Added an OffHeapAllocator for ColumnEncoder. Although working, it is slightly
slower due to the additional copying required to heap when moving to region.
However, the difference is small so if that copying were avoided then it
will certainly be faster (as manual tests have shown) since it can avoid
initializing the memory to zeros like Java has to do.