Addressing #SNAP-1425
Wrapping the first foldable expression containing literals into
DynamicFoldableExpression so that code can be generated at that level
only once per execution in the .init method instead of applying the
expression on a constant for every row during a scan.

Added nested expression verification code that validates the query
result with caching.