Fixed an issue with complex type handling after switching to direct buffers
Issue was that the absolute position of individual fields was being passed in by
generated code to be written in the "offsets+sizes" portion. But this will
cause trouble if the buffer if reallocated to expand it due to being full
during encoding. This did not cause trouble with heap buffers since their
absolute position is the same as relative position in byte[]
(+ fixed Platform.BYTE_ARRAY_OFFSET) but started failing with direct buffers.

Added a detailed comment to ColumnEncoder.initializeComplexType with details
on the encoding layout. Also added some more assertions in the encoding code,
and did some general cleaned up of the code.