correcting an issue with ColumnBatchCreator closing encoder
Some AQP tests failing with IndexOutOfBoundsException due to too large
an allocation is due to ColumnInsertExec closing the encoders at the
end in processNext(). This is invoked by the callback model of sample
table inserts (using ColumnBatchCreator) to start rows in order
to initialize the encoders (which get closed immediately), so now
check for any insertions before closing encoders.