Snap 1304 new (#577)
Implementation of Snapshot Isolation for snappy data.

For Design Please look at : https://docs.google.com/document/d/1bxwSTtk9mW6_qaoubtMsNwm7LGAzBrJA0yTENwIh5uw/edit#heading=h.2h7zw4ru3nw7

The Snapshot is represented by a CopyOnWriteHashMap of RVV information for all regions present in the cache. Each query starts implicitly a transaction with isolation level SNAPSHOT which will store reference to the snapshot when tx is begun.
For ColumnTable, the rollover i.e. flushing of row buffer to column store is a transactional operation.
The rollover process works as follows:

Start a transaction with SNAPSHOT isolation
Iterate over the row buffer to create a batch
Insert into the column store
Delete all the iterated entries
Record the versions related to insert/destroys to snapshot atomically
For ColumnTable, the iteration on buffer and column store is done in a single transaction. It takes a snapshot, which will either see all or none of the changes for any other tx.(e.g roll over of row buffer to column store as shown above)

For smart connector mode, we ensure that for multiple ResultSetIterator same snapshot is used. This is done using procedure to communicate txids across multiple queries.(query on row buffer and column store).

As the SNAPSHOT implementation is per partition, in this implementation, we only guarantee snapshot isolation for any changes in a partition. The work for global snapshot is planned for future and will be on similar lines.