Make a copy of HashMap resident objects for column tables
A previous optimization avoided making a copy of non-primitives coming
from column tables into HashMap (both join and aggregate). This causes
trouble with eviction for both heap and off-heap. For former it causes
the value to hold on to the entire column data inhibiting its overflow.
For latter it can cause crash since there is no handle to the off-heap
buffer at this layer. Even if there was, then it would cause the same
issue of inhibiting entire column's eviction.

Now the copy is made only for row tables since those will hold on to
the row which is much smaller (and eviction for row tables is not
  recommended in any case)

UMM changes:

- increased reserved space to be 1.5X of critical heap so that latter is not hit.
- increased upper limit of minimum eviction to 1G (i.e. eviction can remove
  upto 1G in one go now)