SNAP-1351 (#581)
Changes proposed in this pull request

When tasks are killed, the task threads are getting interrupted. This causes issues with snappy because snappy may be writing to an oplog and it generates a DiskAccessException. This DAE ends up closing the underlying regions.

Snappy threads cannot be interrupted.
Patch testing

Tried out the scenario mentioned in the bug. Works fine. Precheckin.
Other PRs

SnappyDataInc/spark-jobserver#9