Statistics fix and inclusion of snappy-connector
+ Added insert/upsert/delete row count in the metrics as it wasn't
showing up until now.

+ Accomodating snappy-connector/jdbc-stream-connector project into the
compilation if cloned inside snappydata parent project much like aqp.
If needed can add it as a submodule so that it gets automatically
checkedout and the jdbc-stream-connector.jar is published as part of the
SnappyData product distribution. Otherwise user would have to explicitly
add this in the libs folder of the distribution during every upgrade.

For now this would suffice.