Patch From @sumwale
This PR combines the fixes for the two issues:

a) the OOME with large aggregates: new property "snappydata.hashAggregateSize" which is expressed as a string with b/k/m/g units. Kishor, when running with this (hopefully checked in when you run) then do not need the separate patch (the one titled 1651). Set the property to something like 4g: "snappydata.hashAggregateSize=4g"

b) SNAP-1651: it now reads spark driver size and estimates the maximum size to use for a partition result allowing eviction till that point. Rishi, I have removed the array of nonEvictingOwners and added a specific check for partition result to indicate that the workaround only works for partition result (because it applies a specific size limit for eviction and not zero eviction). Also have applied the same to the heap byte[] created at the end because that too can fail with LME frequently.