Support for setting scheduler pools using the set command (#700)
## Changes proposed in this pull request

Added a new scheduler pool in addition to the default pool - lowlatency with FAIR scheduling. The default pool will be used if no pool is set. For now, both the pools have equal weight i.e. same number of cores will be allocated to these pools. If there are no queries of a pool, the cores of that pool are consumed by the jobs of the other pools. A pool can be set using the set command: 

set spark.scheduler.pool=lowlatency;

or using 

sparkContext.setLocalProperty("spark.scheduler.pool", "lowlatency")

We may not document the second way because if the pool does not exist, the API may create a new pool. 

For the set command, we are explicitly throwing an exception if the pool does not exist. 

Right now, we are not providing any property to create new pools during run time. One can do that by updating the conf/fairscheduler.xml before starting the lead node which is the spark way. 

Secondly, This PR has increased plan cache size from 300 to 3000 and has made it configurable. Since SnappySession is part of CachedKey, cached query plans are not getting shared across sessions. The concurrency is badly hit because size of cache is only 300 and multiple concurrent sessions evict each other's plans. For now, increasing the cache size. See SNAP-1605 for more details. 

## Patch testing

Manual. Precheckin.