Disallow update of partitioning/key columns
Else it causes incorrect results without any exception to user in update.
Also added a test to verify the same in ColumnMutableTest.

Fixed the indexes in smart connector reads. Now the delta buffers may not
be present so cannot rely on order of results to match with union all query
so get the base table column index from the results and pass the same
in generated code for getColumnLob.

Revert the change to try and "fix" SNAP-1851 by closing connection in finally block.
That change instead causes lot more failures in SplitSnappyClusterDUnitTest (which
    is a major problem but will be tracked in SNAP-1851)