[Snap 1833] Fix the synchronization problem with sc.addJar() (#728)
SparkContext.addJar is not job specific. For two concurrent jobs, the first spark job which gets scheduled adds jar files for two Snappy jobs. At SnappyExecutor side we keep a map of App-Name to jar path.
But in this case, 2nd jar entry overwrites the first and 1st jar is lost forever.

To fix this issue

Each job will set all dependent jars in job local variable. Those jars will not be added in SparkContext jar lists(sc.addJar). Instead, we will manually fetch those jars from the urls in SnappyExecutor.
The above urls will make a URLClassloader for the duration of jobs. This particular class loader will be set in thread context class loader.
Change the SnappyExecutor to keep a cache of classloader. This cache is required so that multiple tasks of a single job won't be creating multiple classloaders.