SNAP-1884 (#778)
## Changes proposed in this pull request

A replicated table is aggregated/exchanged and the resulting plan is saved as a temporary table. 
This temporary table when joined with SnappyHashJoin, it does not return any rows. It is because 
this temporary table is seen as a replicated table and the aggregation/exchange is not materialized. 
Changed the HashJoinExec to materialize the exchange by adding the correct shuffle dependencies.

## Patch testing

Added a test provided by @ymahajan. Precheckin.