[Snap 1979] MemoryManagerStats for capturing different stats for UMM (#814)
Added MemoryManagerStats for capturing different stats for UMM.
Split mode will not have these stats as GemFireXD cache will not be available. A wrapper class is added for split mode. When cache is initialized it will set the stats object in this wrapper class.
This will avoid null check through out SnappyUnifiedMemoryManager for split and embedded mode.