SNAP:1976: set tx isolation changes (#813)
* changes to set isolation level
Allow operations on row and column tables if isolation level is set to to other than NONE and autocommit is true (query routing is enabled). If autocommit is false, query routing will be disabled and transactions on row tables will are supported. Queries on column tables will error out when query routing is disabled.