Fix for SNAP-1985 (#823)
## Changes proposed in this pull request

Issue is with merging of two deltas. While merging, we create a decoder and initialize it properties based on the decoders that are being merged. But encoder.initialize is not called and stringMap field of the DictionaryEncoder remains uninitialized and hence a null pointer exception. This code calls the initialize passing a minimum size.
Also allows "withHeader=false" for dictionary encoding 

Fix from @sumwale 

## Patch testing

Enabled the test and precheckin.