[SNAP-2020] track in-progress insert size to avoid data skew
With many concurrent inserts/partitions on a node, significant data skew in inserts
was still observed (on machines with large number of cores like 32) due to same
smallest bucket being chosen by multiple partitions.

This change now tracks the in-progress size for bucket and adds that to determine smallest bucket.