[SNAP-2072][SNAP-2073] fix external connectors; support VIEWs (#887)
This commit fixes primarily two issues:

1. External connectors not working in smart connector mode as reported by Jags since the
required libraries may not be available in the embedded cluster. This happens because
the BaseRelation is attempted to be resolved in both "CREATE EXTERNAL TABLE" and
"DROP TABLE". Now resolve all required information (schema, inbuilt or not) at the
driver connector JVM and send that in the procedure calls for external providers.

2. Support for VIEW, VIEW...USING (temporary, global and persistent) in the parser.
   Also cleaned up a few things in parser.

## Changes proposed in this pull request

- fix external connectors in smart connector mode; resolve everything required
  on the source connector and never try to resolve in embedded mode because it
  may not have requisite libraries available
- added VIEW and VIEW...USING support to parser (createView, createTempViewUsing)
- refactored IF NOT EXISTS, IF EXISTS, clauses as separate rules
- replaced remaining uses of trueFn/falseFn by a cleaner push(true)/push(false)

Make SnappySharedState really shared between SnappySessions.

The SparkSession uses a builder approach i.e. getOrCreate() SparkSession
which is a singleton or use newSession for new instances that inherit
SharedState but SnappySession uses cleaner approach of constructor where
the SharedState does not get shared. So override the sharedState
with a global initialized and cleaned after first invokeServices call.

- add tokenization for projection expressions too
- fixing a failure in IndexTest.tpch test due to code generation error in DynamicExpression
  due to sub-expression elimination with a cast(Literal) expression being substituted for
  a cast(ParamLiteral); now skipping sub-expression elimination in DynamicExpression;
  got exposed after the previous change to add tokenization for projection expressions
- fixing TokenizationTest as per current changes to tokenize projections

- adding tests for views:
  - tests for all of temporary views, global temporary views and persisted views
  - fix drop of global temporary view


## Patch testing

precheckin, manual test for external cassandra connector in smart connector mode