[SNAP-2084] Handled dropStorageMemoryForObject in DefaultMemoryManager. (#892)
* [SNAP-2084] Handled dropStorageMemoryForObject in DefaultMemoryManager.
We keep a map of memory used against each region. When we drop the region we release the total memory accumulated. This is not supported in DefaultMemoryManager( which gets used in local & connector mode). Hence the problem.
While acquiring memory for table it uses Spark's UMM , but does not release that memory when the table gets dropped.