[SNAP-2118] allow reading previous variable length value again (#929)
The issue is for the case when variable length value at same position is read multiple times
in Uncompressed encoding. This can happen if aggregation is going on, for example, on
join result where each row on LHS matches multiple on RHS. So then same column value from
LHS can be read multiple times for the aggregate. The code in 1.0.0 release would start
reading all offsets from the start of variable length encoded value.

## Changes proposed in this pull request

- added a "lastSize" member in UncompressedDecoderBase that stores the previously read
  size of variable width value so that cursor can be rewound back to read previously
  returned value again
- added test with perf check for this in ColumnCacheBenchmark
- when testing noticed that escaped LIKE patterns like with '\_' will not convert to StartsWith
  Contains etc if possible so added a LikeEscapeSimplification optimizer rule that also handles
  conversion of backslash escaped wildcard characters
- convert LIKE patterns to StartsWith/EqualTo if possible in predicate pushdown using above
  rule so that those can be optimized using MIN/MAX stats if possible
- added new test suite LikeEscapeSimplificationSuite for above rule