[SNAP-1840] fix NPE in CollectAggregateExec
Issue due to DeserializeToObjectExec using child.output inside the RDD closure
causing child to be serialized all the way to executors (the child plan as such is not
  used otherwise). The CollectAggregateExec was unable to deal with that because it
obtained the output from the basePlan which was deliberately marked as transient.

Now keeping the output separately as non-transient value in CollectAggregateExec.

Added the tests from @ymahajan in https://github.com/SnappyDataInc/snappydata/pull/748