[SNAP-2180] Explicitly initialized memoryMap on UMM start. (#951)
Explicitly initialized memoryMap on UMM start.
When executor goes away we store the UMM stats in another memory manager which is not associated with SparkEnv. Once executors come back we transfer the memory consumption stats to UMM. But this is done in first memory access. Till that time Pulse shows zero memory consumed in UI.

Added a init method to initialize UMM once SparkEnv is ready.

* Fixed test assumptions.