Disabling smokePerf yet again
For the same reasons as mentioned in edd647b1bf7aef9f93ef371305c4d62650c33e1e:

The smokePerf does not work correctly rather leaves around dangling leader/server/locator
processes, so removing it for now till it is fixed.

Long tests like these should not be part of regular precheckin in any case
and should be run with a separate flag when required.