Disabling smokePerf yet again
For the same reasons as mentioned in 7b00b19f3e3ec67c8f2140e072418d09fb62a4f3:

The smokePerf does not work correctly rather leaves around dangling leader/server/locator
processes, so removing it for now till it is fixed.

Long tests like these should not be part of regular precheckin in any case
and should be run with a separate flag when required.