[SNAP-2205] stats row does not have typeId header
Stats row layout is missing any header at the start so its not possible to determine
whether it is compressed or not in the generic case. However changing stats row format
is quite a lot of work due to backward compatibility issues (changing type of on-disk
    and over-the-wire object) with little gain. Instead just check that codecId is one
of the acceptable ones while for the specific case of stats row those can never be hit
because negative of codecId which is written are -1, -2, -3 resolve to 0xfffffff...
which should never happen since nullCount fields are non-nullable in the UnsafeRow
so bitset cannot have 'ff' kind of patterns.