[SNAP-2348] Invalid decompress call on stats row. (#1044)
* [SNAP-2348] Invalid decompress call on stats row.
Even if delta stats rows are not compressed ColumnFormatValue considered it as a compressed value and tries to decompress.

Stats rows are written as UnsafeRow. Unsafe row irrespective of nullability keeps null bits.
So if 100 columns are there stats row will contain 100 * 3 and UnsafeRow will contain
300 null bits plus bits required for 8 bytes word alignment. Hence setting unused columns
as null does not really saves much.

These null bits are set based on platform endianness. SD ColumnFormatValue
assumes LITTLE_ENDIAN bytes. However, Unsafe row itself does not force any endianness
and picks up from the platform. Hence a lower byte of the bit set can be represented as
[11111110]. The 0th bit is for the batch count which can never be null.
This causes SD ColumnFormatValue to understand stats row as a compressed byte
array( as the first int is 1, hence -1 which after taking a negative
equals to 1 i.e LZ4 compression codec id ).
Hence setting each 3rd bit( null count stats) with a not null flag. This will never cause
the word to be read as the negative number.