Spark compatibility (#994)
Adding Spark SQL tests to SnappyData repo. Some tests have been extended. Some tests are running as it is. This commit adds a new module spark-compatibility which contain mostly the test classes and some helper classes. 
Corresponding fixes at the snappy layer is also added. 
Right now there are around 1250 tests from the suite which are passing & and 143 are failing due to different reasons for which analysis is yet to be done. Hence not including this module in pre-check-in. Once all tests start passing will add them to pre-check-in.